Add Google Map to the Map Screen
Implemented a Google Map view in the MapScreen.kt file. The map is initialized with a test location in Lausanne and a zoom level of 10. A marker has also been added at the test location with a title and snippet. This is a preliminary setup and will be updated to follow the user's position and provide some wrappers to display activities with ease