Map with icon and uses the dataclass for activities
I spent a lot of time trying to implement live position in the map but it doesn't work for now.

I created a data class for the activities on the map and you can see an example on the map in lausanne.

This part is not following an MVVM architecture for now and this will be changed in the following days.